+++ apps/mobile/src/App.tsx
@@ -1,0 +1,59 @@
+import { useRef } from 'react';
+
+import { StatusBar } from 'expo-status-bar';
+import { StyleSheet } from 'react-native';
+import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
+import { WebView } from 'react-native-webview';
+
+import { useWebViewHandlers } from './hooks/useWebViewHandlers';
+import { injectedJavaScript } from './lib/injectedJavaScript';
+
+const WEB_APP_URL = process.env.EXPO_PUBLIC_WEB_URL;
+
+const App = () => {
+  const webViewRef = useRef<WebView>(null);
+
+  const { handleShouldStartLoadWithRequest, handleWebViewError } = useWebViewHandlers({
+    webViewRef,
+    webAppUrl: WEB_APP_URL,
+  });
+
+  return (
+    <SafeAreaProvider>
+      <SafeAreaView style={styles.container} edges={['top', 'left', 'right']}>
+        <WebView
+          ref={webViewRef}
+          source={{ uri: WEB_APP_URL }}
+          style={styles.webview}
+          onShouldStartLoadWithRequest={handleShouldStartLoadWithRequest}
+          onError={handleWebViewError}
+          onHttpError={handleWebViewError}
+          injectedJavaScript={injectedJavaScript}
+          javaScriptEnabled
+          domStorageEnabled
+          startInLoadingState
+          scalesPageToFit
+          keyboardDisplayRequiresUserAction={false}
+          mixedContentMode="compatibility"
+          thirdPartyCookiesEnabled
+          sharedCookiesEnabled
+          incognito={false}
+          cacheEnabled
+        />
+        <StatusBar style="auto" />
+      </SafeAreaView>
+    </SafeAreaProvider>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  webview: {
+    flex: 1,
+    backgroundColor: 'transparent',
+  },
+});
+
+export default App;
